+++ crates/renzora_runtime/src/plugins.rs
@@ -79,6 +79,7 @@ pub(crate) fn add_runtime_plugins(app: &mut App) {
     app.add_plugins(renzora_skybox::SkyboxPlugin::default());
     #[cfg(feature = "solari")]
     app.add_plugins(renzora_solari::SolariPlugin::default());
+    app.add_plugins(renzora_spline::SplinePlugin::default());
     #[cfg(feature = "sprite_anim")]
     app.add_plugins(renzora_sprite_anim::SpriteImagesPlugin::default());
     #[cfg(feature = "ssao")]